+++ lib/google_api/configuration.rb
@@ -1,13 +1,23 @@
 module GoogleApi
   class Configuration
 
-    def initialize(config = {})
+    DEFAULT = {
+      client_id: nil,
+      client_secret: nil,
+      client_developer_email: nil,
+      client_cert_file: nil,
+      key_secret: 'notasecret',
+      redirect_uri: nil,
+    }
+
+    def initialize(config, use_default = true)
+
+      if use_default
+        config = DEFAULT.merge(config)
+      end
+
       config.each do |key, value|
         eval <<-METHOD
-          def #{key}=(value)
-            @#{key} = value
-          end
-
           def #{key}(value = nil, &block)
             if block_given?
               @#{key}.instance_eval(&block)
@@ -18,6 +28,10 @@ module GoogleApi
             end
 
             self.#{key} = value
+          end
+
+          def #{key}=(value)
+            @#{key} = value
           end
         METHOD