+++ lib/google_api/ga/ga.rb
@@ -2,13 +2,6 @@ module GoogleApi
   module Ga
 
     CONFIGURATION = {
-      client_id: nil,
-      client_secret: nil,
-      client_developer_email: nil,
-      client_cert_file: nil,
-      key_secret: 'notasecret',
-      redirect_uri: nil,
-
       cache: GoogleApi::Cache.new
     }
 
@@ -23,16 +16,11 @@ module GoogleApi
     autoload :Goal,        'google_api/ga/management/goal'
     autoload :Segment,     'google_api/ga/management/segment'
 
-    # Helper
-    autoload :Helper, 'google_api/ga/helper'
-
     # Data
     autoload :Data,       'google_api/ga/data'
     autoload :DataDsl,    'google_api/ga/data/data_dsl'
     autoload :FiltersDsl, 'google_api/ga/data/filters_dsl'
     autoload :SegmentDsl, 'google_api/ga/data/segment_dsl'
-
-    extend GoogleApi::Ga::Helper
 
     @@id = 0