+++ lib/google_api/shorten/url/url.rb
@@ -1,0 +1,11 @@
+module GoogleApi
+  module Shorten
+    class Url
+      
+      def _session
+        Session
+      end
+
+    end
+  end
+end